+++ python/fnllm/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+- Handle race conditions via error suppression in Azure Blob-Storage cache.
 ### Security
 ```